requirements.txt
altgraph==0.17.2
asgiref==3.5.0
Django==4.0.3
django-ckeditor==6.3.2
django-js-asset==2.0.0
future==0.18.2
pefile==2021.9.3
Pillow==9.1.0
pyinstaller==4.7
pyinstaller-hooks-contrib==2021.4
pywin32-ctypes==0.2.0
sqlparse==0.4.2
tzdata==2021.5
